Refactor packages and introduce package-utils
