+++ packages/kyc-oracle/KycOracle.js
@@ -7,6 +7,7 @@ const { getChallenge, verifyEthSignature } = etherpass;
 const sqlite3 = require("sqlite3");
 const sqlite = require("sqlite");
 const bodyParser = require("body-parser");
+const { colonyIOCors } = require("../package-utils");
 
 class KycOracle {
   /**
@@ -41,24 +42,7 @@ class KycOracle {
 
     this.app = express();
 
-    this.app.use(function (req, res, next) {
-      const origin = req.get("origin");
-
-      const colonyRegex = /.*colony\.io/;
-      const colonyMatches = colonyRegex.exec(origin);
-
-      const localRegex = /http:\/\/(127(\.\d+){1,3}|[0:]+1|localhost)/;
-
-      const localMatches = localRegex.exec(origin);
-
-      if (colonyMatches) {
-        res.header("Access-Control-Allow-Origin", colonyMatches[0]);
-      } else if (localMatches) {
-        res.header("Access-Control-Allow-Origin", "*");
-      }
-      res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept, Synaps-Session-Id");
-      next();
-    });
+    this.app.use(colonyIOCors);
 
     this.app.use(bodyParser.json());
