+++ packages/kyc-oracle/bin/index.js
@@ -10,12 +10,13 @@ const { argv } = require("yargs")
   .option("privateKey", { string: true })
   .option("whitelistAddress", { string: true })
   .option("apiKey", { string: true })
-  .option("adminAddress", { string: true });
+  .option("adminAddress", { string: true })
+  .option("providerAddress", { type: "array", default: [] });
 
 const KycOracle = require("../KycOracle");
 const TruffleLoader = require("../TruffleLoader").default;
 
-const { adminAddress, privateKey, whitelistAddress, apiKey, network, providerPort, providerAddress, dbPath, port } = argv;
+const { adminAddress, privateKey, whitelistAddress, apiKey, network, localProviderPort, localProviderAddress, providerAddress, dbPath, port } = argv;
 const supportedInfuraNetworks = ["mainnet"];
 
 if ((!adminAddress && !privateKey) || !whitelistAddress || !apiKey) {
@@ -34,8 +35,10 @@ if (network) {
     process.exit();
   }
   provider = new ethers.providers.InfuraProvider(network);
+} else if (providerAddress.length === 0) {
+  provider = new ethers.providers.JsonRpcProvider(`http://${localProviderAddress || "localhost"}:${localProviderPort || "8545"}`);
 } else {
-  provider = new ethers.providers.JsonRpcProvider(`http://${providerAddress || "localhost"}:${providerPort || "8545"}`);
+  provider = new ethers.providers.JsonRpcProvider(providerAddress[0]);
 }
 
 const client = new KycOracle({ privateKey, adminAddress, apiKey, loader, provider, dbPath, port });
