+++ packages/package-utils/adapters/console.js
@@ -1,0 +1,19 @@
+class ConsoleAdapter {
+  constructor(label) {
+    if (label) {
+      this.label = `${label}: `;
+    } else {
+      this.label = "";
+    }
+  }
+
+  async log(output) {
+    console.log(this.label, output);
+  }
+
+  async error(output) {
+    console.log(this.label, output);
+  }
+}
+
+export default ConsoleAdapter;
