+++ packages/package-utils/index.js
@@ -1,0 +1,5 @@
+exports.colonyIOCors = require("./colonyIOCors").default;
+exports.updateGasEstimate = require("./updateGasEstimate").default;
+exports.DiscordAdapter = require("./adapters/discord").default;
+exports.SlackAdapter = require("./adapters/slack").default;
+exports.ConsoleAdapter = require("./adapters/console").default;
