+++ packages/package-utils/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "package-utils",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "discord.js": "^12.2.0",
+    "ethers": "^5.5.1",
+    "request-promise": "^4.2.6",
+    "slack": "^11.0.2"
+  }
+}
