+++ packages/reputation-miner/ReputationMinerClient.js
@@ -3,8 +3,7 @@ import apicache from 'apicache'
 const ethers = require("ethers");
 const express = require("express");
 const path = require('path');
-const request = require('request-promise');
-const ConsoleAdapter = require('./adapters/console').default;
+const { ConsoleAdapter, updateGasEstimate } = require('../package-utils');
 
 const ReputationMiner = require("./ReputationMiner");
 
@@ -65,7 +64,7 @@ class ReputationMinerClient {
     }
 
     if (typeof this._adapter === "undefined" ) {
-      this._adapter = ConsoleAdapter;
+      this._adapter = new ConsoleAdapter();
     }
 
     if (this._oracle) {
@@ -284,50 +283,6 @@ class ReputationMinerClient {
     }
   }
 
-  async updateGasEstimate(_type) {
-    let type = _type;
-    const options = {
-      headers: {
-        'User-Agent': 'Request-Promise'
-      },
-      json: true // Automatically parses the JSON string in the response
-    };
-    let defaultGasPrice;
-    let factor;
-
-    if (this.chainId === 100){
-      options.uri = "https://blockscout.com/xdai/mainnet/api/v1/gas-price-oracle";
-      defaultGasPrice = ethers.utils.hexlify(1000000000);
-      factor = 1;
-      // This oracle presents the information slightly differently from ethgasstation.
-      if (_type === "safeLow") {
-        type = "slow";
-      }
-    } else if (this.chainId === 1) {
-      options.uri = "https://ethgasstation.info/json/ethgasAPI.json";
-      defaultGasPrice = ethers.utils.hexlify(20000000000);
-      factor = 10;
-    } else {
-      this._adapter.error(`Error during gas estimation: unknown chainid ${this.chainId}`);
-      this._miner.gasPrice = ethers.utils.hexlify(20000000000);
-      return;
-    }
-
-    // Get latest from whichever oracle
-    try {
-      const gasEstimates = await request(options);
-
-      if (gasEstimates[type]){
-        this._miner.gasPrice = ethers.utils.hexlify(gasEstimates[type] / factor * 1e9);
-      } else {
-        this._miner.gasPrice = defaultGasPrice;
-      }
-    } catch (err) {
-      this._adapter.error(`Error during gas estimation: ${err}`);
-      this._miner.gasPrice = defaultGasPrice;
-    }
-  }
-
   /**
    * Navigate through the mining process logic used when the client is in auto mode.
    * Up to 12 submissions of our current proposed Hash/nLeaves/JRH are made at the earliest block possible
@@ -420,7 +375,7 @@ class ReputationMinerClient {
           if (canSubmit) {
             this._adapter.log("⏰ Looks like it's time to submit an entry to the current cycle");
             this.submissionIndex += 1;
-            await this.updateGasEstimate('average');
+            this.gasPrice = await updateGasEstimate("average", this.chainId, this._adapter);
             await this.submitEntry(entryIndex);
           }
         }
@@ -465,7 +420,7 @@ class ReputationMinerClient {
               return;
             }
           }
-          await this.updateGasEstimate('average');
+          this.gasPrice = await updateGasEstimate("average", this.chainId, this._adapter);
           await repCycle.invalidateHash(round, oppIndex, {"gasPrice": this._miner.gasPrice});
           this.endDoBlockChecks();
           return;
@@ -483,7 +438,7 @@ class ReputationMinerClient {
           );
           if (responsePossible) {
             // If so, invalidate them.
-            await this.updateGasEstimate('average');
+            this.gasPrice = await updateGasEstimate("average", this.chainId, this._adapter);
             await repCycle.invalidateHash(round, oppIndex, {"gasPrice": this._miner.gasPrice});
             this.endDoBlockChecks();
             return;
@@ -496,7 +451,7 @@ class ReputationMinerClient {
         if (submission.jrhNLeaves.eq(0)) {
           const responsePossible = await repCycle.getResponsePossible(disputeStages.CONFIRM_JRH, entry.lastResponseTimestamp);
           if (responsePossible){
-            await this.updateGasEstimate('fast');
+            this.gasPrice = await updateGasEstimate("fast", this.chainId, this._adapter);
             await this._miner.confirmJustificationRootHash();
           }
         // 2. Are we in the middle of a binary search?
@@ -508,7 +463,7 @@ class ReputationMinerClient {
           if (oppEntry.challengeStepCompleted.gte(entry.challengeStepCompleted)) {
             const responsePossible = await repCycle.getResponsePossible(disputeStages.BINARY_SEARCH_RESPONSE, entry.lastResponseTimestamp);
             if (responsePossible){
-              await this.updateGasEstimate('fast');
+            this.gasPrice = await updateGasEstimate("fast", this.chainId, this._adapter);
               await this._miner.respondToBinarySearchForChallenge();
             }
           }
@@ -522,7 +477,7 @@ class ReputationMinerClient {
         {
           const responsePossible = await repCycle.getResponsePossible(disputeStages.BINARY_SEARCH_CONFIRM, entry.lastResponseTimestamp);
           if (responsePossible){
-            await this.updateGasEstimate('fast');
+            this.gasPrice = await updateGasEstimate("fast", this.chainId, this._adapter);
             await this._miner.confirmBinarySearchResult();
           }
         // 4. Is the binary search confirmed, and we need to respond to challenge?
@@ -535,7 +490,7 @@ class ReputationMinerClient {
         {
           const responsePossible = await repCycle.getResponsePossible(disputeStages.RESPOND_TO_CHALLENGE, entry.lastResponseTimestamp);
           if (responsePossible){
-            await this.updateGasEstimate('fast');
+            this.gasPrice = await updateGasEstimate("fast", this.chainId, this._adapter);
             await this._miner.respondToChallenge();
           }
         }
@@ -694,7 +649,7 @@ class ReputationMinerClient {
     const [round] = await this._miner.getMySubmissionRoundAndIndex();
     if (round && round.gte(0)) {
       const gasEstimate = await repCycle.estimateGas.confirmNewHash(round);
-      await this.updateGasEstimate('average');
+      this.gasPrice = await updateGasEstimate("average", this.chainId, this._adapter);
 
       const confirmNewHashTx = await repCycle.confirmNewHash(round, { gasLimit: gasEstimate, gasPrice: this._miner.gasPrice });
       this._adapter.log(`⛏️ Transaction waiting to be mined ${confirmNewHashTx.hash}`);
