--- packages/reputation-miner/adapters/console.js
@@ -1,11 +1,0 @@
-const ConsoleAdapter = {
-  async log(output) {
-    console.log(output);
-  },
-
-  async error(output){
-    console.log(output);
-  }
-}
-
-export default ConsoleAdapter;
